Configure permanent redirects with htaccess
This was misconfigured.  In fact, redirection wasn't even turned on.
Now I've turned it on and added some checks to avoid redirect loops.
Everything that's not a file or directory will result in a permanent
redirect to the site root.  Hopefully this should fix any SEO penalties
we get because of cached links from google leading to blank or bad
not-found pages.

Once we've got more than just a single page of content I'll make this a
bit less heavy and create a 404 page.