Showcase the latest blog post content on the front page
This highlights the most recent blog post with an intro and the entire
content of the post and the author's name signed at the end in a
different coloured block, styled to make it stand out like a snippet
pasted into the front page.  The other blog posts, starting with the
next most recent are then listed underneath as before.

This puts fresh weekly content right on the front page!