Fix spacing in blog list
The loop added newlines before AND after every line of markdown meaning
instead of:

 * foo
 * bar
 * baz

you get:

 * foo

 * bar

 * baz

So it rendered each list item inside its own paragraph which is not only
semantically wrong, it also made the list look pretty weird on the site,
with the items spread far apart.  I tried escaping the newline at the
end of the markdown list item, but that didn't seem to work, and putting
the whole loop on one line made the code look ugly, so I rewrote it as
HTML, which is slightly more verbose, but gives a bit more flexibility
with the final formatting.