Add plugin newlines_to_spaces to fix missing spaces in meta tags
The default strip_newlines filter used to write the title and excerpt in
the meta tags replaces newlines with empty string, meaning if you didn't
leave a trailing space at the end of each line in your paragraph it'll
join the words at the end and start of line into one.

I added a new filter newlines_to_spaces that works the same as the old
one, but replaces them to spaces, and now use that in the meta tags.  I
also filter the result through the strip filter, to make sure you don't
end up with extra spaces at the start or end of the string.