Add plugin for Bible reference links
The links were long and easily generatable, so I wrote a thing to
generated them.  It takes input like this:

    John 4:27-38

and generates output like this:

    [John 4:27-38](https://www.biblegateway.com/passage/?search=John+4%3A27-38&version=WEB)

which saves a lot of typing and a lot of copy-pasting.

It's called "bible".  Call it like this:

    {%bible John 4:27-38%}

It uses the CGI package to do the percent-encoding in the URL, but I
don't think that'll be an issue.