Add 404 error page and turn off forced redirects
When this site started and only had one page, I redirected all missing
files (there were lots because I deleted all the PHP stuff) to the index
page, so that instead of Apache 404 pages, Google would get the index
page and we'd finally get into the results list.  I made the redirects
permanent to reinforce the face that these pages don't exist any more.

Now these permanent redirects are causing problems because when we post
new things the browser remembers the permanent redirect and sends you to
the front page.  I've removed the redirects from the htaccess file and
replaced it with settings for a custom 404 page.  I wrote a 404 page
with info explaining the error and advising action like other pages they
might want to check.  I also set this page as the error page for 403
errors (so far mostly caused when navigating to a directory without an
index file, like /posts/2017/03) because although it's not a 404, the
issue isn't really permission denied, it's "I haven't written an index
generator yet", so it's like 404, at least until I make one.