Temporarily remove reference to the Forum
I should have done this right at the start of the Summer, but better
late than never: It's misleading to write we're meeting at the Forum in
big letters on the website front page when in Summer we actually meet in
parks.  Since new people are finding us via the website, I'm temporarily
removing the reference to the Forum until we actually start meeting
there again.