Add Placeholders.js to show form field placeholders in older browsers
Lots of Windows browsers don't show the placeholder texts and it's bad
for accessibility.

I was considering adding labels above the input fields but that would
add some design complications, especially for mobile, where space is
limited.  Luckily I found a JavaScript polyfill for the HTML5
placeholder attribute, which seems to do a pretty good job of simulating
placeholders in older browsers even as far back as IE6.

All you have to do is download the library (about 4kb) and include it in
your page.  Let's see!