Fix date formatting for custom post dates
These had month and day specified with no year, which in older Jekyll
versions made these be in the current year, making them always appear
newer than the newest posts, and in the new Jekyll version these are
treated as invalid and the posts are not shown at all!  This is
technically better but then the posts are gone, glad I spotted this
subtle formatting issue, all posts should now appear in the correct
order that they were written in!