Update Google Fonts preload code
Checking gstatic results in a 404 error so I dropped that and added this
new preload code that references the same googleapis.com domain that the
font URL is referencing.  Hope it'll be good enough.  Presumably the
gstatic URL wouldn't give 404 when it has the actual font reference in
the URL, but without it I'm not sure it's helping performance giving
ourselves a 404 error automatically.