Add field labels for the signup form
This has been a long-time accessibility and usability issue on the site
for some years and the main impediment to it so far was that I really
liked the way it looked without them and struggled to find a way to make
this fit.  I took the extra effort now to make it work with different
layouts so that on laptop and desktop they all fit in one line and for
tablets, mobiles etc. they rearrange vertically.