Align submit button to right for medium devices
When the width is small enough that we enlarge the form to better use
the space, there's an ugly gap on the right-hand side.  Floating the
button to the right solves this, I'm not sure why it requires adding the
top-margin again but it works.  The float needs to be removed again for
the palm/phone layout when everything is vertical, because otherwise it
slips out of the form element… a cool overlay effect, but not what I
wanted for this flat style.