Change all Bible references from WEB to NIV
The original reason behind using the WEB version was that its licence is
very permissive about usage but in practice the version most of us own a
copy of and read and quote from is NIV, or similar to that.  I'm making
this the default now and considering that all these links have a
reference to an online source where you can seen which version it's from
and also that we are only quoting small parts inside the context of our
own writing I think this counts as "fair use" compared to e.g. posting a
copy of the full book.