Improve markup and formatting for the retreat post
Probably I was in a bit too much of a hurry when I published this back
in the day, some blank lines missing after quotes turned everything
after it into part of the quote too which looked really weird.  Fixed
that and a few other things too while I'm here.